+++ gl-core/cartridges/Fallmanager/data/caseObj.tsx
